+++ tests/spec/js/schema.d.ts
@@ -1711,17 +1711,20 @@ export interface FullRequestParams extends Omit<RequestInit, "body"> {
   body?: unknown;
   /** base url */
   baseUrl?: string;
+  /** request cancellation token */
+  cancelToken?: CancelToken;
 }
 export declare type RequestParams = Omit<FullRequestParams, "body" | "method" | "query" | "path">;
 interface ApiConfig<SecurityDataType = unknown> {
   baseUrl?: string;
-  baseApiParams?: Omit<RequestParams, "baseUrl">;
+  baseApiParams?: Omit<RequestParams, "baseUrl" | "cancelToken" | "signal">;
   securityWorker?: (securityData: SecurityDataType) => RequestParams | void;
 }
 interface HttpResponse<D extends unknown, E extends unknown = unknown> extends Response {
   data: D;
   error: E;
 }
+declare type CancelToken = Symbol | string | number;
 export declare enum ContentType {
   Json = "application/json",
   FormData = "multipart/form-data",
@@ -1731,6 +1734,7 @@ export declare class HttpClient<SecurityDataType = unknown> {
   baseUrl: string;
   private securityData;
   private securityWorker;
+  private abortControllers;
   private baseApiParams;
   constructor(apiConfig?: ApiConfig<SecurityDataType>);
   setSecurityData: (data: SecurityDataType) => void;
@@ -1739,6 +1743,8 @@ export declare class HttpClient<SecurityDataType = unknown> {
   protected addQueryParams(rawQuery?: QueryParamsType): string;
   private contentFormatters;
   private mergeRequestParams;
+  private createAbortSignal;
+  abortRequest: (cancelToken: CancelToken) => void;
   request: <T = any, E = any>({
     body,
     secure,
@@ -1747,6 +1753,7 @@ export declare class HttpClient<SecurityDataType = unknown> {
     query,
     format,
     baseUrl,
+    cancelToken,
     ...params
   }: FullRequestParams) => Promise<HttpResponse<T, E>>;
 }